[refactor-payment-integrations] Created payment integrations folder and rename paypalContributions to paypalPaymentAPI
