+++ assets/helpers/paymentIntegrations/stripeCheckout.js
